+++ feature/signup/src/main/java/com/kusitms/connectdog/signup/screen/common/CompleteSignUpScreen.kt
@@ -1,4 +1,4 @@
-package com.kusitms.connectdog.signup.screen
+package com.kusitms.connectdog.signup.screen.common
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
