+++ feature/signup/src/main/java/com/kusitms/connectdog/signup/screen/common/RegisterPasswordScreen.kt
@@ -1,11 +1,12 @@
-package com.kusitms.connectdog.signup.screen
+package com.kusitms.connectdog.signup.screen.common
 
 import android.annotation.SuppressLint
-import android.widget.Toast
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -16,7 +17,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
@@ -26,40 +27,39 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.kusitms.connectdog.core.designsystem.R
 import com.kusitms.connectdog.core.designsystem.component.ConnectDogNormalButton
-import com.kusitms.connectdog.core.designsystem.component.ConnectDogTextFieldWithButton
+import com.kusitms.connectdog.core.designsystem.component.ConnectDogTextField
 import com.kusitms.connectdog.core.designsystem.component.ConnectDogTopAppBar
 import com.kusitms.connectdog.core.designsystem.component.TopAppBarNavigationType
 import com.kusitms.connectdog.core.designsystem.theme.ConnectDogTheme
+import com.kusitms.connectdog.core.designsystem.theme.Gray3
 import com.kusitms.connectdog.core.designsystem.theme.Orange_40
 import com.kusitms.connectdog.core.designsystem.theme.PetOrange
 import com.kusitms.connectdog.core.util.UserType
-import com.kusitms.connectdog.signup.viewmodel.RegisterEmailViewModel
+import com.kusitms.connectdog.signup.viewmodel.RegisterPasswordViewModel
 import com.kusitms.connectdog.signup.viewmodel.SignUpViewModel
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
-fun RegisterEmailScreen(
+fun RegisterPasswordScreen(
     onBackClick: () -> Unit,
+    onNavigateToVolunteerProfile: (UserType) -> Unit,
+    onNavigateToIntermediatorProfile: () -> Unit,
     userType: UserType,
-    onNavigateToRegisterPassword: (UserType) -> Unit,
+    imeHeight: Int,
     signUpViewModel: SignUpViewModel,
-    viewModel: RegisterEmailViewModel = hiltViewModel(),
-    imeHeight: Int
+    viewModel: RegisterPasswordViewModel = hiltViewModel()
 ) {
     val focusManager = LocalFocusManager.current
     val interactionSource = remember { MutableInteractionSource() }
-    val context = LocalContext.current
-    val isValidEmail by viewModel.isValidEmail.collectAsState()
-    val isEmailVerified by viewModel.isEmailVerified.collectAsState()
-    val isEmailDuplicated by viewModel.isEmailDuplicated.collectAsState()
+    val isValidPassword by viewModel.isValidPassword.collectAsState()
+    val isValidConfirmPassword by viewModel.isValidConfirmPassword.collectAsState()
 
     Scaffold(
         topBar = {
             ConnectDogTopAppBar(
                 titleRes = when (userType) {
-                    UserType.SOCIAL_VOLUNTEER -> R.string.volunteer_signup
-                    UserType.NORMAL_VOLUNTEER -> R.string.volunteer_signup
                     UserType.INTERMEDIATOR -> R.string.intermediator_signup
+                    else -> R.string.volunteer_signup
                 },
                 navigationType = TopAppBarNavigationType.BACK,
                 navigationIconContentDescription = "Navigation icon",
@@ -69,6 +69,8 @@ fun RegisterEmailScreen(
     ) {
         Column(
             modifier = Modifier
+                .fillMaxSize()
+                .background(Color.White)
                 .padding(horizontal = 20.dp)
                 .clickable(
                     onClick = { focusManager.clearFocus() },
@@ -78,71 +80,62 @@ fun RegisterEmailScreen(
         ) {
             Spacer(modifier = Modifier.height(80.dp))
             Text(
-                text = "로그인에 사용할\n이메일을 입력해주세요",
+                text = "로그인에 사용할\n비밀번호를 입력해주세요",
                 fontSize = 20.sp,
                 fontWeight = FontWeight.Bold
             )
             Spacer(modifier = Modifier.height(40.dp))
-            ConnectDogTextFieldWithButton(
-                text = viewModel.email,
-                width = 62,
-                height = 27,
-                textFieldLabel = "이메일",
-                placeholder = "이메일 입력",
-                buttonLabel = "인증 요청",
-                isError = (isValidEmail ?: false) || (isEmailDuplicated ?: false),
+            ConnectDogTextField(
+                text = viewModel.password,
+                label = "비밀번호",
+                placeholder = "비밀번호 입력",
+                keyboardType = KeyboardType.Password,
+                isError = isValidPassword ?: false,
                 onTextChanged = {
-                    viewModel.updateEmail(it)
-                    viewModel.updateEmailValidity()
-                },
-                onClick = {
-                    if (isValidEmail == true) {
-                        Toast.makeText(context, "유효한 이메일을 입력해주세요", Toast.LENGTH_SHORT).show()
-                    } else if (isEmailDuplicated == true) {
-                        Toast.makeText(context, "이미 가입된 이메일입니다.", Toast.LENGTH_SHORT).show()
-                    } else {
-                        viewModel.postEmail()
-                        Toast.makeText(context, "이메일을 전송했습니다", Toast.LENGTH_SHORT).show()
-                    }
-                },
-                padding = 5
+                    viewModel.updatePassword(it)
+                    viewModel.checkPasswordValidity(it)
+                }
             )
             Spacer(modifier = Modifier.height(12.dp))
-            ConnectDogTextFieldWithButton(
-                text = viewModel.certificationNumber,
-                width = 62,
-                height = 27,
-                textFieldLabel = "인증 번호",
-                placeholder = "숫자 6자리",
-                buttonLabel = "인증 확인",
-                keyboardType = KeyboardType.Text,
-                onTextChanged = { viewModel.updateCertificationNumber(it) },
-                onClick = {
-                    if (it.isNotEmpty()) {
-                        viewModel.checkCertificationNumber()
-                    } else {
-                        Toast.makeText(context, "인증번호를 입력해주세요", Toast.LENGTH_SHORT).show()
-                    }
-                },
-                padding = 5
+            ConnectDogTextField(
+                text = viewModel.confirmPassword,
+                label = "비밀번호 확인",
+                placeholder = "비밀번호 확인",
+                keyboardType = KeyboardType.Password,
+                isError = isValidConfirmPassword ?: false,
+                onTextChanged = {
+                    viewModel.updateConfirmPassword(it)
+                    viewModel.checkConfirmPasswordValidity(it)
+                }
+            )
+            Spacer(modifier = Modifier.height(4.dp))
+            Text(
+                text = "영문+숫자 10자 이상 또는 영문+숫자+특수기호 8자 이상",
+                modifier = Modifier.padding(start = 8.dp),
+                fontSize = 11.sp,
+                color = Gray3
             )
             Spacer(modifier = Modifier.weight(1f))
             ConnectDogNormalButton(
                 content = "다음",
-                color = if (isEmailVerified) {
+                modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .height(56.dp),
+                color = if (isValidPassword == false && isValidConfirmPassword == false) {
                     PetOrange
                 } else {
                     Orange_40
                 },
                 onClick = {
-                    if (isEmailVerified) {
-                        signUpViewModel.updateEmail(viewModel.email)
-                        onNavigateToRegisterPassword(userType)
+                    if (isValidPassword == false && isValidConfirmPassword == false) {
+                        signUpViewModel.updatePassword(viewModel.password)
+                        when (userType) {
+                            UserType.INTERMEDIATOR -> onNavigateToIntermediatorProfile()
+                            else -> onNavigateToVolunteerProfile(userType)
+                        }
                     }
-                },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(56.dp)
+                }
             )
             Spacer(modifier = Modifier.height((imeHeight + 32).dp))
         }
@@ -151,14 +144,8 @@ fun RegisterEmailScreen(
 
 @Preview
 @Composable
-private fun test() {
+private fun Preview() {
     ConnectDogTheme {
-        RegisterEmailScreen(
-            onBackClick = {},
-            userType = UserType.NORMAL_VOLUNTEER,
-            onNavigateToRegisterPassword = {},
-            hiltViewModel(),
-            imeHeight = 10
-        )
+        RegisterPasswordScreen({}, {}, {}, UserType.INTERMEDIATOR, 10, hiltViewModel())
     }
 }
